test(int): :rotating_light: use globalSetup instead
