+++ tests/api/package.json
@@ -13,11 +13,10 @@
     "@types/supertest": "2.0.12",
     "ant-stack": "workspace:*",
     "body-parser": "1.20.2",
-    "consola": "3.2.3",
     "cors": "2.8.5",
+    "cross-fetch": "4.0.0",
     "esbuild": "0.19.1",
     "express": "4.18.2",
-    "supertest": "6.3.3",
     "vitest": "0.34.1"
   }
 }
