+++ tests/api/tsconfig.json
@@ -1,8 +1,4 @@
 {
   "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": { "@fixtures": ["./src/fixtures/index.ts"] }
-  },
   "include": [".", "src"]
 }
